+++ examples/simple_fit_binned.py
@@ -23,7 +23,7 @@ data_binned = data.to_binned(obs_binned)  # could also be a hist object
 
 nll = zfit.loss.BinnedNLL(model=gauss_binned, data=data_binned)
 
-minimizer = zfit.minimize.Minuit()
+minimizer = zfit.minimize.Minuit(gradient="zfit")
 result = minimizer.minimize(nll)
 
 param_errors = result.hesse()